Moar refactoring
